requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.1.0
httpx==0.28.1
supabase==2.15.0
requests==2.32.3
pydantic==2.11.1 